major version bump
Bumped the "big" major version number. We had some backward compatibility
changes that required a major bump, but we elected to change the "big"
version because that better reflects the maturity of the library. The
only reason the previous version was 0.1.0.0 was because this is newly
open-sourced. But the private library from which it is derived was at
4.4.0.0 when we decided to open source it.