feat: Updated `events` to catch invalid `diffMerge` cases. ...
This commit changes the type of `events` such that it will refuse
to provide a `Diff` for a particular target when the corresponding
`diffMerge` might be unsafe or result in an inconsistency. It had
previously been up to the client to determine whether a `diffMerge` is
valid, which is bad API design, and furthermore the complexity of making
this determination was greater than I originally believed, making even
worse API design for the client to have to figure this out.