no-op: Fix documentation. ...
This commit fixes the haddock documentation in `Data.CRDT.EventFold.Monad`
by correcting the links into the main module. I had previously thought
that haddock was smart enough to understand aliased qualified module
imports, but it is not. You have to use the fully qualified "true"
name of the module.